Rename SampleSet* tests to SampleType*
