+++ src/org/labkey/test/params/experiment/SampleTypeDefinition.java
@@ -14,7 +14,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import static org.labkey.test.tests.SampleSetTest.SAMPLE_TYPE_DOMAIN_KIND;
+import static org.labkey.test.tests.SampleTypeTest.SAMPLE_TYPE_DOMAIN_KIND;
 
 /**
  * Defines a Sample Type. Suitable for use with UI and API helpers.
